pass non-script specs
