+++ test/execute/script-spec.ts
@@ -3,7 +3,7 @@ import { execFileSync } from 'child_process'
 import { } from 'mocha'
 
 describe('script', () => {
-  const hc_bin = 'lib/cli/hc.js'
+  const hc_bin = 'src/cli/hc.ts'
   let title: string
 
   beforeEach(function () {
